implement fixes requested for Connect Certification (2023-10-02) (#180)
* improve subscription management in Connect deployment scripts
